remove priority_work_queue since we're not using it
